Move logging code to seperate files
Avoids unnecessary recompilation
