+++ src/btop_menu.cpp
@@ -24,12 +24,13 @@ tab-size = 4
 #include <cmath>
 #include <filesystem>
 
-#include "btop_menu.hpp"
-#include "btop_tools.hpp"
 #include "btop_config.hpp"
-#include "btop_theme.hpp"
 #include "btop_draw.hpp"
+#include "btop_log.hpp"
+#include "btop_menu.hpp"
 #include "btop_shared.hpp"
+#include "btop_theme.hpp"
+#include "btop_tools.hpp"
 
 using robin_hood::unordered_flat_map;
 using std::array;
