+++ src/freebsd/btop_collect.cpp
@@ -60,6 +60,7 @@ tab-size = 4
 #include <memory>
 
 #include "../btop_config.hpp"
+#include "../btop_log.hpp"
 #include "../btop_shared.hpp"
 #include "../btop_tools.hpp"
