+++ src/linux/btop_collect.cpp
@@ -40,8 +40,9 @@ tab-size = 4
 	#include <pwd.h>
 #endif
 
-#include "../btop_shared.hpp"
 #include "../btop_config.hpp"
+#include "../btop_log.hpp"
+#include "../btop_shared.hpp"
 #include "../btop_tools.hpp"
 
 using std::clamp;
