+++ src/osx/btop_collect.cpp
@@ -53,6 +53,7 @@ tab-size = 4
 #include <string>
 
 #include "../btop_config.hpp"
+#include "../btop_log.hpp"
 #include "../btop_shared.hpp"
 #include "../btop_tools.hpp"
